+++ src/test/java/dev/openfeature/sdk/HookContextTest.java
@@ -29,52 +29,4 @@ class HookContextTest {
                     "The before stage MUST run before flag resolution occurs. It accepts a hook context (required) and hook hints (optional) as parameters. It has no return value.")
     @Test
     void not_applicable_for_dynamic_context() {}
-
-    @Test
-    void shouldCreateHookContextWithHookData() {
-        HookContext<String> context = HookContext.<String>builder()
-                .flagKey("test-flag")
-                .type(FlagValueType.STRING)
-                .defaultValue("default")
-                .ctx(new ImmutableContext())
-                .build();
-        HookData hookData = HookData.create();
-        hookData.set("test", "value");
-
-        HookContextWithData contextWithData = new HookContextWithData(context, hookData);
-
-        assertNotNull(contextWithData.getHookData());
-        assertEquals("value", contextWithData.getHookData().get("test"));
-    }
-
-    @Test
-    void shouldCreateHookContextWithoutHookData() {
-        HookContext<String> context = HookContext.<String>builder()
-                .flagKey("test-flag")
-                .type(FlagValueType.STRING)
-                .defaultValue("default")
-                .ctx(new ImmutableContext())
-                .build();
-
-        assertNull(context.getHookData());
-    }
-
-    @Test
-    void shouldCreateHookContextWithHookDataUsingWith() {
-        HookContext<String> originalContext = HookContext.<String>builder()
-                .flagKey("test-flag")
-                .type(FlagValueType.STRING)
-                .defaultValue("default")
-                .ctx(new ImmutableContext())
-                .build();
-
-        HookData hookData = HookData.create();
-        hookData.set("timing", System.currentTimeMillis());
-
-        HookContext<String> contextWithHookData = new HookContextWithData(originalContext, hookData);
-
-        assertNull(originalContext.getHookData());
-        assertNotNull(contextWithHookData.getHookData());
-        assertNotNull(contextWithHookData.getHookData().get("timing"));
-    }
 }
